feat(zsh): rename install_oh_my_zsh.sh to install_oh_my_zsh.zsh
Renamed the install_oh_my_zsh.sh script to install_oh_my_zsh.zsh to better reflect the fact that it's a Zsh script. This change was made to improve the consistency of file extensions across the project.

BREAKING CHANGE: This change renames the file extension of the install script from .sh to .zsh, which may affect any scripts or tooling that rely on the previous file extension.